IGNITE-23689 Add tests for standalone compute
